+++ src/main/java/com/github/yangricardo/api_spring_boot/shared/configurations/security/SecurityConfiguration.java
@@ -49,8 +49,8 @@ public class SecurityConfiguration extends WebSecurityConfigurerAdapter {
     @Override
     protected void configure(HttpSecurity http) throws Exception {
         http.authorizeRequests().antMatchers(HttpMethod.POST, "/auth/**").permitAll().antMatchers("/values/**")
-                .permitAll().anyRequest().authenticated().and().csrf().disable().sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
+                .permitAll().antMatchers("/http-request-sample/**").permitAll().anyRequest().authenticated().and()
+                .csrf().disable().sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS).and()
                 .addFilterBefore(this.tokenAuthenticationFilterService, UsernamePasswordAuthenticationFilter.class);
         ;
     }