+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Standardization of project for `com.github.yangricardo` base package
 - Lombok classes definition improvement
 - Serializing operations of Java classes to JSON objects notation
+- Sample of external http rest api call on [HttpRequestSampleController](src/main/java/com/github/yangricardo/api_spring_boot/modules/http_request_sample/controllers/HttpRequestSampleController.java)
+- enable of Spring WebFlux
 
 ## 2021.11.18